trait_diet as JSON
omnivore eats both meat and plants (like people do, usually). However, it may over-complicate so we could delete or just not use/implement it now. This uses an array also. again the first and last line define the array. I don't know if we need an array or if it can use the file itself with raw data in it.